Fixing a security issue.
Previously, a bad server could impersonate a peer with 0 effort:

(rpk means random pulic exchange key)

A sends a rpk to B
This message is intercepted by the server; The server proceeds to
generate a new random rpk, just as B would do; The server sends this key
back to A.
A doesn't know what happend, everything looks normal, except it isn't.
A proceeds to send an encrypted message with their shared key.
the imposter B can sucessfully decrypt and read it.

Bad, very bad.

To mitigate this, now A encrypts the message twice, the first time with the
shared secret of A(sk) and B(pk), the second one with the shared secret
of A(rpk(sk)) and B(rpk(pk)).
The bad server can sucessfully decrypt the first layer, since they
generated a rpk and so have the private part of it. The second layer
is not possible to decrypt, if you don't have B's private key, which the
server thankfully doesn't have access to :).

Problem solved?